+++ test/functional/member/friendships_controller_test.rb
@@ -24,10 +24,10 @@ class Member::FriendshipsControllerTest < ActionController::TestCase
       
       should "send an info message" do
         assert_equal 1, @chavez.inbox.messages.count
-        assert_equal I18n.t("tog_social.friendships.member.mail.add_friend.subject", :user_name => @berlusconi.profile.full_name), @chavez.inbox.messages.first.subject
+        assert_equal I18n.t("tog_social.friendships.member.mail.request_friend.subject", :user_name => @berlusconi.profile.full_name), @chavez.inbox.messages.first.subject
       end
 
-      should_set_the_flash_to I18n.t("tog_social.friendships.member.friend.added", :friend_name => "chavez")
+      should_set_the_flash_to I18n.t("tog_social.friendships.member.friend.request", :friend_name => "chavez")
     end
     
     context "on POST to :confirm_friend" do