+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Edge
 
 * Replaced own sharing code by acts_as_shareable plugin. Please, install acts_as_shareable and it's migration before running plugins migrations.
 * Acts_as_state_machine code replaced by AASM gem
+* Better messages for bidirectional friendships
 
 0.5.4
 ----